core/auth/local.py
"""Local auth provider — JWT + bcrypt."""

from __future__ import annotations

import uuid
from datetime import UTC, datetime, timedelta

import bcrypt
from jose import JWTError, jwt

from core.auth.base import AbstractAuthProvider, User


class LocalAuthProvider(AbstractAuthProvider):
    def __init__(self, secret: str, algorithm: str = "HS256", expire_minutes: int = 60) -> None:
        self._secret = secret
        self._algorithm = algorithm
        self._expire_minutes = expire_minutes

    async def authenticate(self, token: str) -> User:
        try:
            payload = jwt.decode(token, self._secret, algorithms=[self._algorithm])
        except JWTError as exc:
            raise ValueError("Invalid or expired token") from exc

        return User(
            id=uuid.UUID(payload["sub"]),
            email=payload["email"],
            tenant_id=uuid.UUID(payload["tenant_id"]),
            roles=payload.get("roles", []),
        )

    async def create_token(self, user: User) -> str:
        now = datetime.now(UTC)
        payload = {
            "sub": str(user.id),
            "email": user.email,
            "tenant_id": str(user.tenant_id),
            "roles": user.roles,
            "iat": now,
            "exp": now + timedelta(minutes=self._expire_minutes),
        }
        return jwt.encode(payload, self._secret, algorithm=self._algorithm)

    async def hash_password(self, password: str) -> str:
        return bcrypt.hashpw(password.encode(), bcrypt.gensalt()).decode()

    async def verify_password(self, password: str, hashed: str) -> bool:
        return bcrypt.checkpw(password.encode(), hashed.encode())
