pyproject.toml
[project]
name = "privai-platform"
version = "0.1.0"
description = "Privacy management platform — Data Classification, DSAR, and Veil"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.2",
    "alembic>=1.14",
    "aiofiles>=24.1",
]

[dependency-groups]
dev = [
    "ruff>=0.9",
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
    "import-linter>=2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "api", "products", "settings.py"]

# ── Ruff ───────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py313"
line-length = 100
extend-exclude = ["*.sky"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"api/dependencies.py" = ["B008"]  # FastAPI Depends() in default args is idiomatic

[tool.ruff.lint.isort]
known-first-party = ["core", "api", "products", "settings"]

# ── Pytest ─────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

# ── Import Linter ──────────────────────────────────────────────────────
[tool.importlinter]
root_packages = ["core", "api", "products"]

[[tool.importlinter.contracts]]
name = "Products never import from each other"
type = "independence"
modules = [
    "products.classify",
    "products.dsar",
    "products.veil",
]

[[tool.importlinter.contracts]]
name = "Products never import from external packages"
type = "forbidden"
source_modules = [
    "products.classify",
    "products.dsar",
    "products.veil",
]
forbidden_modules = []

[[tool.importlinter.contracts]]
name = "Core never imports from api or products"
type = "forbidden"
source_modules = ["core"]
forbidden_modules = ["api", "products"]

[[tool.importlinter.contracts]]
name = "Community API modules use only community implementations"
type = "forbidden"
source_modules = [
    "api.main",
    "api.dependencies",
    "api.routes.dsar",
    "api.middleware",
]
forbidden_modules = [

    "core.tenancy.multi",


]
